requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.16
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
kombu==5.4.2
prompt-toolkit==3.0.48
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==5.2.0
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
